Add CreatedAt and LastUpdatedAt fields to the transactions model
