+++ HousingSearchApi.Tests/HousingSearchApi.Tests.csproj
@@ -45,7 +45,7 @@
     <PackageReference Include="Docker.DotNet" Version="3.125.4" />
     <PackageReference Include="FluentAssertions" Version="5.10.3" />
     <PackageReference Include="Hackney.Core.ElasticSearch" Version="1.45.0" />
-    <PackageReference Include="Hackney.Shared.HousingSearch" Version="0.18.0" />
+    <PackageReference Include="Hackney.Shared.HousingSearch" Version="0.19.0" />
     <PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" Version="3.0.0" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.4" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.3" />
