+++ HousingSearchApi/V1/Factories/ResponseFactory.cs
@@ -30,7 +30,9 @@ namespace HousingSearchApi.V1.Factories
                 domain.HousingBenefitAmount,
                 domain.Address,
                 domain.Sender.ToDomain(),
-                domain.Fund);
+                domain.Fund,
+                domain.CreatedAt,
+                domain.LastUpdatedAt);
         }
 
         public static List<TransactionResponse> ToResponse(this IEnumerable<Transaction> domainList)
